REF: move get_op_result_name out of ops.__init__
